+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 [Full diff](https://github.com/sider/runners/compare/0.44.1...HEAD)
 
 - **GolangCI-Lint** Set explicitly `GO111MODULE=auto` [#2129](https://github.com/sider/runners/pull/2129)
+- **Flake8** Add `parallel` option [#2130](https://github.com/sider/runners/pull/2130)
 
 ## 0.44.1